move match_{params,results} methods to wasmi
