+++ crates/wasmi/src/func/error.rs
@@ -1,4 +1,3 @@
-use crate::core::FuncTypeError as CoreFuncTypeError;
 use core::{
     error::Error,
     fmt::{self, Display},
@@ -17,18 +16,6 @@ pub enum FuncError {
     MismatchingResultType,
     /// Specified an incorrect number of results.
     MismatchingResultLen,
-}
-
-impl From<CoreFuncTypeError> for FuncError {
-    fn from(error: CoreFuncTypeError) -> Self {
-        match error {
-            CoreFuncTypeError::MismatchingParameterType => Self::MismatchingParameterType,
-            CoreFuncTypeError::MismatchingParameterLen => Self::MismatchingParameterLen,
-            CoreFuncTypeError::MismatchingResultType => Self::MismatchingResultType,
-            CoreFuncTypeError::MismatchingResultLen => Self::MismatchingResultLen,
-            error => panic!("unsupported error: {error:?}"),
-        }
-    }
 }
 
 impl Error for FuncError {}
