+++ crates/wasmi/src/func/func_type.rs
@@ -1,12 +1,90 @@
 use crate::{
-    core::{DynamicallyTyped, FuncType as CoreFuncType, ValType},
+    core::{FuncType as CoreFuncType, ValType},
     func::FuncError,
     Val,
 };
 
+/// Types that are dynamically typed, such as [`ValType`].
+pub trait DynamicallyTyped {
+    /// Returns the [`ValType`] of `self`.
+    fn ty(&self) -> ValType;
+}
+
+impl DynamicallyTyped for ValType {
+    fn ty(&self) -> ValType {
+        *self
+    }
+}
+
 impl DynamicallyTyped for Val {
     fn ty(&self) -> ValType {
         self.ty()
+    }
+}
+
+/// Extension methods for [`FuncType`].
+trait FuncTypeExt {
+    /// Returns `Ok` if the number and types of items in `params` matches as expected by the [`FuncType`].
+    ///
+    /// # Errors
+    ///
+    /// - If the number of items in `params` does not match the number of parameters of the function type.
+    /// - If any type of an item in `params` does not match the expected type of the function type.
+    fn match_params<T>(&self, params: &[T]) -> Result<(), FuncError>
+    where
+        T: DynamicallyTyped;
+
+    /// Returns `Ok` if the number and types of items in `results` matches as expected by the [`FuncType`].
+    ///
+    /// # Note
+    ///
+    /// Only checks types if `check_type` is set to `true`.
+    ///
+    /// # Errors
+    ///
+    /// - If the number of items in `results` does not match the number of results of the function type.
+    /// - If any type of an item in `results` does not match the expected type of the function type.
+    fn match_results<T>(&self, results: &[T], check_type: bool) -> Result<(), FuncError>
+    where
+        T: DynamicallyTyped;
+}
+
+impl FuncTypeExt for CoreFuncType {
+    fn match_params<T>(&self, params: &[T]) -> Result<(), FuncError>
+    where
+        T: DynamicallyTyped,
+    {
+        if self.params().len() != params.len() {
+            return Err(FuncError::MismatchingParameterLen);
+        }
+        if self
+            .params()
+            .iter()
+            .copied()
+            .ne(params.iter().map(<T as DynamicallyTyped>::ty))
+        {
+            return Err(FuncError::MismatchingParameterType);
+        }
+        Ok(())
+    }
+
+    fn match_results<T>(&self, results: &[T], check_type: bool) -> Result<(), FuncError>
+    where
+        T: DynamicallyTyped,
+    {
+        if self.results().len() != results.len() {
+            return Err(FuncError::MismatchingResultLen);
+        }
+        if check_type
+            && self
+                .results()
+                .iter()
+                .copied()
+                .ne(results.iter().map(<T as DynamicallyTyped>::ty))
+        {
+            return Err(FuncError::MismatchingResultType);
+        }
+        Ok(())
     }
 }
